Convert int matrix to float

+++ src/Matrix.java
@@ -2,18 +2,18 @@ import java.util.Arrays;
 import java.util.Scanner;
 
 public class Matrix{
-    int[][] matrix;
+    float[][] matrix;
     int size;
     public Matrix(int size){
         this.size = size;
-        matrix = new int[size][size];
+        matrix = new float[size][size];
     }
 
-    public void SetValue(int row, int column, int value){
+    public void SetValue(int row, int column, float value){
         matrix[row][column] = value;
     }
 
-    public int GetValue(int row, int column){
+    public float GetValue(int row, int column){
         return matrix[row][column];
     }
